pyproject.toml
[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.dependencies]

python = "^3.12"
python-dotenv = "^1"
litestar = { extras = ["standard"], version = "^2" }
litestar-granian = "^0"
litestar-asyncpg = "^0"
advanced-alchemy = "^0.20"
msgspec = "^0.18.6"
alembic = "1.15.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


psycopg = "^3.2.6"
[project]
name = "litestar-crud-app"
version = "0.1.0"
description = "LiteStar CRUD API for User entity with PostgreSQL"
readme = "README.md"
authors = [
    {name = "Lukmanov Nikitam", email = "lukmanov.nikita06@mail.ru"},
]
requires-python = "^3.12"
dynamic = ["dependencies"]

[project.scripts]
start = "src.app:run_server"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

